Separate US and World data services into different packages
