+++ src/main/java/com/royware/corona/dashboard/config/ApplicationConfig.java
@@ -23,7 +23,7 @@ import org.springframework.scheduling.concurrent.ConcurrentTaskScheduler;
 import org.springframework.web.client.RestTemplate;
 
 import com.royware.corona.dashboard.interfaces.data.ExternalDataService;
-import com.royware.corona.dashboard.services.data.ExternalDataServiceWorldImpl;
+import com.royware.corona.dashboard.services.data.world.ExternalDataServiceWorldImpl;
 
 @Configuration
 @PropertySource({"classpath:application-${ENVIRONMENT}.properties"})
