+++ src/main/java/com/royware/corona/dashboard/services/data/cache/CacheActionsUnitedStatesImpl.java
@@ -1,4 +1,4 @@
-package com.royware.corona.dashboard.services.data;
+package com.royware.corona.dashboard.services.data.cache;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -15,6 +15,7 @@ import com.royware.corona.dashboard.interfaces.data.ExternalDataService;
 import com.royware.corona.dashboard.interfaces.data.ICacheActions;
 import com.royware.corona.dashboard.interfaces.model.CanonicalCaseDeathData;
 import com.royware.corona.dashboard.model.data.us.UnitedStatesData;
+import com.royware.corona.dashboard.services.data.world.ExternalDataServiceWorldImpl;
 
 @Component("ca_us")
 public class CacheActionsUnitedStatesImpl implements ICacheActions {
