+++ src/main/java/com/royware/corona/dashboard/services/data/us/ExternalDataServiceUSAImpl.java
@@ -1,4 +1,4 @@
-package com.royware.corona.dashboard.services.data;
+package com.royware.corona.dashboard.services.data.us;
 
 import java.util.List;
 
@@ -14,6 +14,7 @@ import com.royware.corona.dashboard.interfaces.data.ExternalDataService;
 import com.royware.corona.dashboard.interfaces.data.ExternalDataServiceFactory;
 import com.royware.corona.dashboard.interfaces.data.IMultiRegionExternalDataService;
 import com.royware.corona.dashboard.model.data.us.UnitedStatesData;
+import com.royware.corona.dashboard.services.data.cache.CacheManagerProvider;
 
 @Component("us")
 public class ExternalDataServiceUSAImpl implements ExternalDataService {
