+++ src/main/java/com/royware/corona/dashboard/services/data/us/MultiRegionListStitcherImpl.java
@@ -1,4 +1,4 @@
-package com.royware.corona.dashboard.services.data;
+package com.royware.corona.dashboard.services.data.us;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
