+++ src/main/java/com/royware/corona/dashboard/services/data/world/ExternalDataServiceSingleCountryImpl.java
@@ -1,4 +1,4 @@
-package com.royware.corona.dashboard.services.data;
+package com.royware.corona.dashboard.services.data.world;
 
 import java.util.ArrayList;
 import java.util.Collections;
